requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
DateTimeRange==1.2.0
discord.py==1.7.3
idna==3.3
isort==5.10.1
jsonmerge==1.8.0
jsonschema==4.13.0
mbstrdecoder==1.1.1
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
typepy==1.3.0
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1
